requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
google-api-python-client==2.109.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
gspread-django==3.0.6
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-allauth==0.57.0
openpyxl==3.1.2
pandas==2.1.3
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
